Added t-tests for Selenium
1. on brine vs bin1, bin2, bin3, bin4.
It'd answer the simple question that if the bins work, but now backed up by numbers.
Then we can pick the best bin (prob same as the boxplot, which is bin 2 and 6).
2. on the treatment train. So bin1 vs bin5, then bin5 vs bin7. Next is bin2 vs bin6.
This is for see if there's an optimal level of treatment train levels, i.e. just 1 bin is good, or 2, or 3.